+++ src/refactoring/components/admin/index.ts
@@ -1,0 +1,3 @@
+export * from './CouponManage';
+export * from './NewProduct';
+export * from './ProductInfo';
