+++ src/refactoring/stores/index.ts
@@ -1,0 +1,2 @@
+export * from './useCouponStore';
+export * from './useProductStore';
